refactor: extract ConnectionCommonMixin to deduplicate sync/async connection code
Moves shared state init, validation, pure helpers (_invalidate_query_handles,
_ensure_connected, _safe_close_socket, _drop_connection), and CAS_INFO
constants into _connection_common.py. Both Connection and AsyncConnection
inherit from ConnectionCommonMixin.

Closes #133
